+++ CHANGELOG.md
@@ -89,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Websearch tools.
 - Small bugfixes.
 - [frontend] - Fix suggestion UI.
+- [frontend] - cosmetic fixes.
 
 ## [0.3.3] - 30.10.2024